text-editor: move keydown handler to hook
